Never exceed original counts
With this commit, the output is equivalent to part of a shuffled list of
the count-repeated list of hashes.  In other words, it is equivalent to
what these steps would produce: (1) expand the original list of hashes
according to their counts, (2) shuffle the list, and (3) take e.g. first
one million lines from the shuffled list.

Prior to this commit, the counts were treated only as probabilities, which
meant that e.g. a line with a count of 1 could occasionally appear more
than once in the output.